+++ project/urls.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.urls import include, path
 
-from url_shortener.views import URLRedirectAPIView
+from url_shortener.views.redirect_view import URLRedirectAPIView
 
 urlpatterns = [
     path("admin/", admin.site.urls),
